remove more circular dependencies
